requirements.txt
torch~=1.12.0
numpy~=1.21.6
pandas~=1.4.2
matplotlib~=3.5.2
scikit-learn~=1.1.1
h5py~=3.7.0
torchsummary~=1.5.1
plotly~=5.8.0
py3Dmol~=1.8.0
seaborn~=0.11.2
umap-learn[plot]
hurry.filesize
scipy~=1.8.1
tensorboard
filprofiler